Conflicts:
	addons/account/i18n/vi.po
	addons/account_bank_statement_import/i18n/vi.po
	addons/account_check_printing/i18n/vi.po
	addons/auth_ldap/i18n/vi.po
	addons/auth_oauth/i18n/vi.po
	addons/base_import/i18n/vi.po
	addons/base_setup/i18n/vi.po
	addons/base_vat/i18n/vi.po
	addons/calendar_sms/i18n/vi.po
	addons/delivery/i18n/vi.po
	addons/event_sale/i18n/vi.po
	addons/google_calendar/i18n/vi.po
	addons/google_drive/i18n/vi.po
	addons/google_spreadsheet/i18n/vi.po
	addons/hr/i18n/vi.po
	addons/hr_expense/i18n/vi.po
	addons/hr_fleet/i18n/vi.po
	addons/hr_holidays/i18n/vi.po
	addons/hr_maintenance/i18n/vi.po
	addons/hr_presence/i18n/vi.po
	addons/hr_recruitment/i18n/vi.po
	addons/hr_skills/i18n/vi.po
	addons/im_livechat/i18n/vi.po
	addons/link_tracker/i18n/vi.po
	addons/mail/i18n/vi.po
	addons/maintenance/i18n/vi.po
	addons/mass_mailing/i18n/vi.po
	addons/mass_mailing_slides/i18n/vi.po
	addons/membership/i18n/vi.po
	addons/mrp/i18n/vi.po
	addons/payment/i18n/vi.po
	addons/payment_payumoney/i18n/vi.po
	addons/payment_sips/i18n/vi.po
	addons/payment_stripe/i18n/vi.po
	addons/payment_transfer/i18n/vi.po
	addons/portal/i18n/vi.po
	addons/pos_cash_rounding/i18n/vi.po
	addons/pos_discount/i18n/vi.po
	addons/pos_restaurant/i18n/vi.po
	addons/project_timesheet_holidays/i18n/vi.po
	addons/rating/i18n/vi.po
	addons/sale/i18n/vi.po
	addons/sale_coupon/i18n/vi.po
	addons/sale_expense/i18n/vi.po
	addons/sale_management/i18n/vi.po
	addons/sale_product_matrix/i18n/vi.po
	addons/sale_quotation_builder/i18n/vi.po
	addons/sale_stock/i18n/vi.po
	addons/sale_timesheet/i18n/vi.po
	addons/sales_team/i18n/vi.po
	addons/stock_landed_costs/i18n/vi.po
	addons/survey/i18n/vi.po
	addons/uom/i18n/vi.po
	addons/utm/i18n/vi.po
	addons/web_tour/i18n/vi.po
	addons/website/i18n/vi.po
	addons/website_crm/i18n/vi.po
	addons/website_forum/i18n/vi.po
	addons/website_rating/i18n/vi.po
	addons/website_sale_delivery/i18n/vi.po
	addons/website_slides_survey/i18n/vi.po

+++ addons/website_sale/models/sale_order.py
@@ -149,12 +149,12 @@ class SaleOrder(models.Model):
 
         try:
             if add_qty:
-                add_qty = float(add_qty)
+                add_qty = int(add_qty)
         except ValueError:
             add_qty = 1
         try:
             if set_qty:
-                set_qty = float(set_qty)
+                set_qty = int(set_qty)
         except ValueError:
             set_qty = 0
         quantity = 0
